# Conflicts:
#	SofaKernel/modules/SofaBaseCollision/CubeModel.cpp
#	SofaKernel/modules/SofaBaseCollision/CubeModel.h
#	SofaKernel/modules/SofaMeshCollision/TriangleModel.h
#	applications/plugins/ColladaSceneLoader/CMakeLists.txt
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/Compiler/pushpack1.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/Exporter.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/IOStream.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/IOSystem.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/Importer.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/LogStream.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/Logger.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/ProgressHandler.hpp
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/ai_assert.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/anim.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/cexport.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/cimport.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/color4.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/color4.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/config.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/defs.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/light.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/material.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/material.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/matrix3x3.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/matrix3x3.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/matrix4x4.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/matrix4x4.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/mesh.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/postprocess.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/quaternion.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/quaternion.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/scene.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/types.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/vector2.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/vector2.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/vector3.h
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/vector3.inl
#	applications/plugins/ColladaSceneLoader/assimp/include/assimp/version.h
#	applications/plugins/ColladaSceneLoader/assimp/lib/x64/Assimp64.dll
#	applications/plugins/ColladaSceneLoader/assimp/lib/x64/assimp.lib
#	modules/SofaExporter/src/SofaExporter/BlenderExporter.inl

+++ applications/plugins/Flexible/BaseJacobian.h
@@ -179,24 +179,24 @@ protected:
 
 
 
-template<class Real1, class Real2,  int Dim1, int Dim2>
+template<class Real1, class Real2,  size_t Dim1, size_t Dim2>
 inline Mat<Dim1, Dim2, Real2> covMN(const Vec<Dim1,Real1>& v1, const Vec<Dim2,Real2>& v2)
 {
     Mat<Dim1, Dim2, Real2> res;
-    for ( unsigned int i = 0; i < Dim1; ++i)
-        for ( unsigned int j = 0; j < Dim2; ++j)
+    for (size_t i = 0; i < Dim1; ++i)
+        for ( size_t j = 0; j < Dim2; ++j)
         {
             res[i][j] = (Real2)v1[i] * v2[j];
         }
     return res;
 }
 
-template<class Real,  int Dim>
+template<class Real,  size_t Dim>
 inline MatSym<Dim, Real> covN(const Vec<Dim,Real>& v)
 {
     MatSym<Dim, Real> res;
-    for ( unsigned int i = 0; i < Dim; ++i)
-        for ( unsigned int j = i; j < Dim; ++j)
+    for (size_t i = 0; i < Dim; ++i)
+        for (size_t j = i; j < Dim; ++j)
         {
             res(i,j) = v[i] * v[j];
         }
